+++ account-keeper-node/src/main/java/com/dwarfeng/acckeeper/node/launcher/Launcher.java
@@ -3,6 +3,7 @@ package com.dwarfeng.acckeeper.node.launcher;
 import com.dwarfeng.acckeeper.node.handler.LauncherSettingHandler;
 import com.dwarfeng.acckeeper.stack.service.CleanQosService;
 import com.dwarfeng.acckeeper.stack.service.ProtectorSupportMaintainService;
+import com.dwarfeng.acckeeper.stack.service.ResetQosService;
 import com.dwarfeng.springterminator.sdk.util.ApplicationUtil;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import org.slf4j.Logger;
@@ -66,6 +67,32 @@ public class Launcher {
                             }
                         },
                         new Date(System.currentTimeMillis() + startCleanupDelay)
+                );
+            }
+
+            // 处理重置处理器的启动选项。
+            ResetQosService resetQosService = ctx.getBean(ResetQosService.class);
+            // 重置处理器是否启动重置服务。
+            long startResetDelay = launcherSettingHandler.getStartResetDelay();
+            if (startResetDelay == 0) {
+                LOGGER.info("立即启动重置服务...");
+                try {
+                    resetQosService.start();
+                } catch (ServiceException e) {
+                    LOGGER.error("无法启动重置服务，异常原因如下", e);
+                }
+            } else if (startResetDelay > 0) {
+                LOGGER.info(startResetDelay + " 毫秒后启动重置服务...");
+                scheduler.schedule(
+                        () -> {
+                            LOGGER.info("启动重置服务...");
+                            try {
+                                resetQosService.start();
+                            } catch (ServiceException e) {
+                                LOGGER.error("无法启动重置服务，异常原因如下", e);
+                            }
+                        },
+                        new Date(System.currentTimeMillis() + startResetDelay)
                 );
             }
         });